Use long rather than int values for MIL constants so that we can accommodate both 64 bit (long) and 32 bit (int) targets.  (When WORDSIZE is 32, the long values stored in IntConst objects are truncated appropriately to match the range of int.)  Includes changes to core lexer (eliminates getSmallInt()); MIL and LC ASTs; and to optimizer, plus misc other pieces.  The types Long and long are not currently used for any other purposes in this code, so searching for those can be a good way to identify parts of the system that are sensitive to WORDSIZE variations (in addition to those parts that depend either directly or indirectly on the value of Type.WORDSIZE).

+++ src/llvm/Int.java
@@ -23,10 +23,10 @@ package llvm;
 public class Int extends Value {
 
   /** The numeric value associated with this Int. */
-  private int num;
+  private long num;
 
   /** Default constructor. */
-  public Int(int num) {
+  public Int(long num) {
     this.num = num;
   }
